+++ module/module.js
@@ -38,7 +38,7 @@ Hooks.once("ready", async () => {
 
     if (foundry.utils.isNewerVersion(currentVersion, lastVersion)) {
         const journal = await fromUuid("Compendium.fvtt-trazzm-homebrew-5e.homebrew-journal-entries.JournalEntry.L9YHsoeODbdhqdKU");
-        const page = journal.pages.contents[journal.pages.contents.length - 1];
+        const page = journal.pages.contents.find(p => p.name === currentVersion);
         journal.sheet.render(true, {pageId: page.id});
         game.settings.set("fvtt-trazzm-homebrew-5e", "lastVersion", currentVersion)
     }