+++ scripts/macros/fix-scene-tokens.js
@@ -1,0 +1,6 @@
+for (let token of canvas.scene.tokens) {
+    await token.update({
+        displayBars: 30,
+        displayName: 30
+    });
+}