+++ changelog.md
@@ -2,8 +2,12 @@
 * Add support for v13 with DND5.x
 
 # 13.5.1
-* Updates for 12.4.21
+* Updates from 12.4.21
 * Work on getting the settings properly configured
 
 # 13.5.2
-* 
+* Updates from 12.4.22
+* Added macros to fix world and scene tokens to show health and name on hover
+* Added Pact of the Blade to fix enchanting issues
+* Fixed Great Weapon Fighting feat
+* Updated settings and module list exports